add opt-out Cargo features to remove the default initialization and ...
panic handler routines. This way we can stop relying on the #[linkage =
"weak"] feature which is unstable and can't be used to override the
these routines from a dependency (i.e. these routines can only be
overridden from the top crate) because of how linker arguments are
ordered by `rustc`.